GS-14503 handle multiple conditions in join (Sagiv), support array[index] in where/join, temp disable exception of JOIN with LITERAL/FUNCTION

+++ xap-extensions/xap-jdbc/src/main/java/com/gigaspaces/jdbc/calcite/pg/PgCalciteTable.java
@@ -99,7 +99,7 @@ public class PgCalciteTable extends AbstractTable {
             String fqn = "public." + name;
             int oid = PgOidGenerator.INSTANCE.oid(fqn);
             ITypeDesc typeDesc = SQLUtil.checkTableExistence(name, space);
-            short idx = 0;
+            long idx = 0;
             for (PropertyInfo property : typeDesc.getProperties()) {
                 SqlTypeName sqlTypeName = GSTable.mapToSqlType(property.getType());
                 PgTypeDescriptor pgType = PgTypeUtils.fromSqlTypeName(sqlTypeName);
